Build warnings revisited (#2318)
* unused vars in bison parser cleanup

* initialization order in TopologicalTorsionGenerator

* unused params in SLN bison

* sln flex unused params

* throwing destructor in TDTWriter

* signed comparison in substructmethods

* unused input param in smiles/smarts bison

* unused ms param in sln bison

* signed comparison in FingerprintGenerator

* store return of fscanf in StructCheckerOptions

* unreferenced var in catch

* uninitialized value in FileParserUtils

* avoid override overload warning in MolDraw2DSVG

* non-final overrides in Validate.h

* unused static var in Avalon

* unused vars in catch blocks

* make AvalonTools avalonSimilarityBits & avalonSSSBits const int

* assert fscanf result in StructCheckerOptions

+++ Code/GraphMol/Wrap/substructmethods.h
@@ -101,7 +101,7 @@ PyObject *helpGetSubstructMatches(T1 &mol, T2 &query,
     matches = SubstructMatch(mol, query, params);
   }
   PyObject *res = PyTuple_New(matches.size());
-  for (int idx = 0; idx < matches.size(); idx++) {
+  for (size_t idx = 0; idx < matches.size(); idx++) {
     PyTuple_SetItem(res, idx, convertMatches(matches[idx]));
   }
   return res;
